make PageBanner generic for all page types
