+++ src/components/Pages/Guide/_Guide.scss
@@ -1,14 +1,5 @@
 $sticky-guide-top: 25px;
 
-.coa-GuidePageBannerImage {
-  object-fit: cover;
-  // This is for https://www.npmjs.com/package/object-fit-images
-  font-family: 'object-fit: cover;';
-
-  height: 32rem;
-  width: 100%;
-}
-
 .coa-GuidePage-header--container {
   width: 57rem;
   margin-left: auto;
